Add functionality to check slave's state frequently
Added new static function `checkSlaveState` in `Slave.h` and its definition in `Slave.cpp` to constantly keep track of the slave state. The function checks if a pump is actually active or has some error. For handling the checks, a new timer `shellyupdate` is set up which triggers after every 15000ms. It helps in notifying errors like "Slave is empty" if the pump is presumably active, but the power is less than the defined minimum, or in case of unclosed contactor for the slave. For defining the power limits, new constants `MIN_SLAVE_POWER` and `MAX_SLAVE_DISABLED` are added to `PINOUT.h`.